backend: implemented functionality to update the teams' scores (instead of resetting the complete database)

+++ web/defaults.py
@@ -2,8 +2,19 @@ from database_models import db
 from database_models import Team
 from database_models import *
 
-db.drop_all()
-db.create_all()
+import argparse
+parser = argparse.ArgumentParser()
+parser.add_argument("--update", help="only update existing scores", action="store_true")
+args = parser.parse_args()
+
+if args.update:
+    print "...only doing an update"
+else:
+	print "...clearing and re-filling DB" 
+	db.drop_all()
+	db.create_all()
+
+
 
 session = getSession()
 
@@ -83,63 +94,64 @@ all_teams = [
 	CompactTeamData("Korea Republic", "KR", "H", 1683, 556, 52500000 / 22.0, 25.50, SPI(73.4, 1.7, 1.0))
 	]
 
-# first step: insert teams into DB to get the IDs
-for team_data in all_teams:
-	team = Team(team_data.name, team_data.country_code)
-	session.add(team)
-# add new rating score types
-session.add(ScoreType("ELO", "The ELO rating known from chess.", long_name="ELO rating"))
-session.add(ScoreType("FIFA", "FIFA ranking points", long_name="FIFA rating"))
-session.add(ScoreType("Value", "Average value of the players in Euro", long_name="&#216; value in &euro;"))
-session.add(ScoreType("Age", "Average age of the team", long_name="&#216; age"))
-session.add(ScoreType("HA", "Home-advantage of the team", long_name="Home-advantage"))
-session.add(ScoreType("Custom", "User-defined custom rating", long_name="Custom rating", hidden=True))
-session.add(ScoreType("SPI Off", "ESPN SPI Offensive Rating", long_name="SPI Offensive Rating"))
-session.add(ScoreType("SPI Def", "ESPN SPI Defensive Rating", long_name="SPI Defensive Rating"))
+if not args.update:
+	# first step: insert teams into DB to get the IDs
+	for team_data in all_teams:
+		team = Team(team_data.name, team_data.country_code)
+		session.add(team)
+	# add new rating score types
+	session.add(ScoreType("ELO", "The ELO rating known from chess.", long_name="ELO rating"))
+	session.add(ScoreType("FIFA", "FIFA ranking points", long_name="FIFA rating"))
+	session.add(ScoreType("Value", "Average value of the players in Euro", long_name="&#216; value in &euro;"))
+	session.add(ScoreType("Age", "Average age of the team", long_name="&#216; age"))
+	session.add(ScoreType("HA", "Home-advantage of the team", long_name="Home-advantage"))
+	session.add(ScoreType("Custom", "User-defined custom rating", long_name="Custom rating", hidden=True))
+	session.add(ScoreType("SPI Off", "ESPN SPI Offensive Rating", long_name="SPI Offensive Rating"))
+	session.add(ScoreType("SPI Def", "ESPN SPI Defensive Rating", long_name="SPI Defensive Rating"))
 
-# add certain custom parameters
-custom_rule_parameter = RuleParameterType("normalization_constant", 10.0)
-session.add(custom_rule_parameter)
+	# add certain custom parameters
+	custom_rule_parameter = RuleParameterType("normalization_constant", 10.0)
+	session.add(custom_rule_parameter)
 
-session.add(RuleParameterType("simulation_run_count", 1))
+	session.add(RuleParameterType("simulation_run_count", 1))
 
-# add default ELO calculation rule
-elo_rule = RuleType("ELO", "Calculation using the ELO score", "elo_binary")
-elo_rule.makeDefaultRule(1.0)
-session.add(elo_rule)
+	# add default ELO calculation rule
+	elo_rule = RuleType("ELO", "Calculation using the ELO score", "elo_binary")
+	elo_rule.makeDefaultRule(1.0)
+	session.add(elo_rule)
 
-spi_rule = RuleType("SPI", "Calculation based on ESPN's Soccer-Power-Index", "spi_binary")
-spi_rule.makeDefaultRule(1.0)
-session.add(spi_rule)
+	spi_rule = RuleType("SPI", "Calculation based on ESPN's Soccer-Power-Index", "spi_binary")
+	spi_rule.makeDefaultRule(1.0)
+	session.add(spi_rule)
 
-fifa_rule = RuleType("FIFA", "Calculation using the FIFA ranking", "fifa_binary")
-fifa_rule.makeDefaultRule(0.5)
-session.add(fifa_rule)
+	fifa_rule = RuleType("FIFA", "Calculation using the FIFA ranking", "fifa_binary")
+	fifa_rule.makeDefaultRule(0.5)
+	session.add(fifa_rule)
 
-value_rule = RuleType("Value", "Calculation based on average player market value", "value_binary")
-value_rule.makeDefaultRule(0.25)
-session.add(value_rule)
+	value_rule = RuleType("Value", "Calculation based on average player market value", "value_binary")
+	value_rule.makeDefaultRule(0.25)
+	session.add(value_rule)
 
-ha_rule = RuleType("HA", "Adjust the win expectancy based on the home-advantage", "homeadvantage_binary", long_name="Home-advantage", is_backref_rule=True)
-ha_rule.makeDefaultRule(1.0)
-session.add(ha_rule)
+	ha_rule = RuleType("HA", "Adjust the win expectancy based on the home-advantage", "homeadvantage_binary", long_name="Home-advantage", is_backref_rule=True)
+	ha_rule.makeDefaultRule(1.0)
+	session.add(ha_rule)
 
-age_rule = RuleType("Age", "Calculation based on average age", "age_binary")
-age_rule.makeDefaultRule(0.0)
-session.add(age_rule)
+	age_rule = RuleType("Age", "Calculation based on average age", "age_binary")
+	age_rule.makeDefaultRule(0.0)
+	session.add(age_rule)
 
-luck_rule = RuleType("Luck", "Each team has the same probability of winning", "luck_binary")
-luck_rule.makeDefaultRule(0.0)
-session.add(luck_rule)
+	luck_rule = RuleType("Luck", "Each team has the same probability of winning", "luck_binary")
+	luck_rule.makeDefaultRule(0.0)
+	session.add(luck_rule)
 
-custom_rule = RuleType("Custom", "Define custom ratings and an own win expectancy function.", "custom_binary", long_name="Custom Rating", needs_custom_ratings=True)
-custom_rule.makeDefaultRule(0.0)
-session.add(custom_rule)
-# add default tournament types
-session.add(TournamentType("1 vs 1", "A simple 1 vs 1 test tournament.", 2, "TwoHandsIcon.png", "1v1"))
-session.add(TournamentType("World Cup", "The standard FIFA World Cup.", 32, "StdLeagueIcon.png", "worldcup", "worldcup_view"))
-# only after comitting will the objects have valid IDs assigned!
-session.commit()
+	custom_rule = RuleType("Custom", "Define custom ratings and an own win expectancy function.", "custom_binary", long_name="Custom Rating", needs_custom_ratings=True)
+	custom_rule.makeDefaultRule(0.0)
+	session.add(custom_rule)
+	# add default tournament types
+	session.add(TournamentType("1 vs 1", "A simple 1 vs 1 test tournament.", 2, "TwoHandsIcon.png", "1v1"))
+	session.add(TournamentType("World Cup", "The standard FIFA World Cup.", 32, "StdLeagueIcon.png", "worldcup", "worldcup_view"))
+	# only after comitting will the objects have valid IDs assigned!
+	session.commit()
 
 # get the objects we just added (now with correct ID)
 elo = session.query(ScoreType).filter_by(name="ELO").first()
@@ -159,17 +171,18 @@ assert custom != None
 assert spi_off != None
 assert spi_def != None
 
-elo_rule.addScoreType(elo, session)
-fifa_rule.addScoreType(fifa, session)
-value_rule.addScoreType(value, session)
-age_rule.addScoreType(age, session)
-ha_rule.addScoreType(ha, session)
+if not args.update:
+	elo_rule.addScoreType(elo, session)
+	fifa_rule.addScoreType(fifa, session)
+	value_rule.addScoreType(value, session)
+	age_rule.addScoreType(age, session)
+	ha_rule.addScoreType(ha, session)
 
-custom_rule.addScoreType(custom, session)
-custom_rule.addParameterType(custom_rule_parameter, session)
+	custom_rule.addScoreType(custom, session)
+	custom_rule.addParameterType(custom_rule_parameter, session)
 
-spi_rule.addScoreType(spi_off, session)
-spi_rule.addScoreType(spi_def, session)
+	spi_rule.addScoreType(spi_off, session)
+	spi_rule.addScoreType(spi_def, session)
 
 
 
@@ -178,14 +191,25 @@ for team_data in all_teams:
 	team = session.query(Team).filter_by(country_code=team_data.country_code).first()
 	assert team != None
 	
-	session.add(Score(elo.id, team.id, team_data.ELO))
-	session.add(Score(fifa.id, team.id, team_data.FIFA))
-	session.add(Score(value.id, team.id, team_data.Value))
-	session.add(Score(age.id, team.id, team_data.Age))
-	session.add(Score(ha.id, team.id, team_data.HA))
-	session.add(Score(spi_off.id, team.id, team_data.spi.off_rating))
-	session.add(Score(spi_def.id, team.id, team_data.spi.def_rating))
-
+	fun = None
+	def add(type, score):
+			session.add(Score(type.id, team.id, score))
+	def update(type, score):
+			session.query(Score).filter_by(type_id=type.id,tournament_id=None,team_id=team.id).first().value = score
+	if not args.update:
+		fun = add
+	else:
+		fun = update
+		
+	fun(elo, team_data.ELO)
+	fun(fifa, team_data.FIFA)
+	fun(value, team_data.Value)
+	fun(age, team_data.Age)
+	fun(ha, team_data.HA)
+	fun(spi_off, team_data.spi.off_rating)
+	fun(spi_def, team_data.spi.def_rating)
+	
+	
 session.commit()
 print "..done"
 cleanupSession()
